refactor(edit-engine): extract Route transaction domains
Move Route point authoring and geometry-only removal into a geometry dispatcher, and move Junction lifecycle, route attachment, cutting, endpoint connection, and disconnection into a topology dispatcher.

Preserve ordered-edit dependencies, explicit Route geometry requirements for Junction movement, split Net identity and Evidence retargeting, exact rejection paths/object IDs, deferred pruning, and connectivity outcome propagation. Post-loop physical-contact normalization and final Route validation remain in executeTransaction.

Validation: pnpm typecheck; pnpm test:local packages/edit-engine/src (246 tests); focused manual-editor routing browser coverage (10 scenarios); git diff --check.

No test files changed because existing transaction and browser routing contracts directly exercise the moved seams.

Test-Impact: no-test-change — behavior-preserving Route/Junction dispatcher extraction covered by 246 package tests and 10 browser scenarios.
